requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Deprecated==1.2.14
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-social-oauth2==2.1.3
ecdsa==0.18.0
future==0.18.3
gunicorn==21.2.0
idna==3.4
joblib==1.3.1
jwcrypto==1.5.0
lightgbm==4.0.0
numpy
oauthlib==3.2.2
packaging==23.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
python-jose==3.3.0
python-whois==0.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
threadpoolctl
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
whitenoise==6.5.0
wrapt==1.15.0
